Refactor BlankLinePhase into subphases
